feat: remove teardown on terraform apply failure
The call to `up.sh` was being combined with `|| down.sh` for the purposes of tearing down the
testnet if a failure occurs on the `terraform apply`. The problem with this is it masks a failure,
because it makes the script return a positive exit code. This has the effect that if you're using
this action and your testnet doesn't spin up correctly, the job you use it in will pass. You really
want to know that things failed at the `terraform apply` stage.

If this job fails, you can clean up the testnet using another stage/job that always runs on failure.
You may even want to keep the testnet around so that you can investigate the cause of the failed
Terraform run. You need to give the caller of the action the opportunity to do that, without
destroying the network for them. Some notes were added to the README for this.