chore: provide more debugging info for callers
There are a bunch of files that are generated by the testnet tool that can be useful to callers of
the action. In particular, they can be used by our nightly run to facilitate uploading logs when the
run fails.

These files get copied to a system-wide location for easy access. They would be accessible from
wherever the files for the custom action get cloned to, but this is something done by GHA and is in
some awkward location.