feat: introduce testnet-id parameter
Give control of the testnet ID to the caller of the action.

This means the caller can retrieve debugging information from S3 based on the name of the testnet
ID, which they have control over, and so information doesn't need to be passed out of the custom
action.

This removes the need for copying debugging information to the system-wide location, which was not
very useful anyway, because it's only accessible within the same job, which isn't very useful.