fix: explicitly declare aws region variable
For some unknown reason, the `Create testnet` step was setting `AWS_DEFAULT_REGION` to empty, even
though that variable wasn't declared as part of that step and it was assigned a value in the
previous step. It seems that an environment variable passed to the process from the parent
environment will override the value of any variables set in the .env file, so `AWS_DEFAULT_REGION`
was being set to an empty, and hence Terraform was then complaining about it. The same thing applied
for the `Destroy testnet` step.