chore: provide `network-contacts-file-name` input
The `testnet-deploy` tool has been extended to upload the network contacts file to S3 on each
deployment. By default, the name of the contacts file will be the same as the environment name. It
has an option to supply a different name for the file, which could be useful for restricted
networks.