chore: use network contacts file name for inv cmd
The `inventory` command for `testnet-deploy` accepts an argument for the network contacts file name
output. By exposing this argument on the print inventory workflow, we can generate a new network
contacts file after a testnet has already been deployed. This is useful if the network contacts file
was not uploaded with the correct name.