feat: provide vm and bootstrap node count inputs
Three new VM inputs are introduced: bootstrap-node-vm-count, node-vm-count, uploader-vm-count. As
the names imply, they control the number of VMs for each of those categories. A deployment now
includes bootstrap nodes and 'normal' nodes, as well as uploader machines. The `bootstrap-peer`
input is removed because two deploy runs are not required.

A fourth input is also added: `bootstrap-node-count`. It controls the number of node services that
run on the bootstrap VMs, allowing that to be changed independently of the number that run on
generic nodes, which is much larger.

The `node-count` input is also made optional here. It did not use a `required` attribute, but it was
being called explicitly in the command definition without checking to see if the variable was set.